DISCORD/requirements.txt
discord
python-valve
steamid
mysql
steam